Apply cleanup-review findings to the SlideBox extraction (#1801)
- Promote SlideBox to the shared ui/compose/components package (it is
  generic in shape and docs; tests move with it) and reword its docs in
  widget vocabulary, demoting EtaStrip mentions to examples.
- Be honest about the ownership guarantee: the single-owner invariant
  covers the widget's own mutations; the hoisted ScrollState remains
  observable, and its docs now say observe-don't-drive.
- Expose SlideBoxState.isSettledAtEnd so the reveal's settled state is
  defined once by the widget; EtaStrip's settle predicate no longer
  re-derives it from raw scroll state.
- Dedup the two regime blocks into one glideRegime helper so the
  cancel-and-join invariant is written once.
- Restore the allocation-free early returns on the nested-scroll hot
  path (PullDelta is now only built while a pull is actively being
  built/unwound, not per pointer delta of every drag).
- Inline the stale justifyIndex local (its justify job moved into the
  widget's first-alignment snap), drop a redundant same-value write
  guard, and drop a dead -1 sentinel from SlideBoxTest.
